render layer drawing improvement
realized my 4 am code was setting the view every object. it no longer does that.

next, make a spriterenderer, and in the inspector, add a way to browse the file path. can even use the texture size parameters to display a preview of the sprite chosen. also try to make sure it can only select sprites within the assets/sprites folder.