Close to basic player movement
just need to figure out oncollision exit stuff.
rn its buggy since i use isgrounded to stop downwards movement when i collide with a floor, but then try to set isgrounded when we no longer collide with an object.

maybe modify manifolds/collision system to allow for objects that are touching but not overlapping.